pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyneople"
version = "1.0.0"
description = "Neople Open API wrapper with MongoDB and PostgreSQL support for data pipelines"
readme = "README.md"
requires-python = ">=3.9"

authors = [
  { name = "ippo252525", email = "ippo252525@gmail.com" }
]

classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

dependencies = [
  "motor>=3.7.0,<4.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "aiohttp>=3.11.14,<4.0.0",
  "python-dotenv>=1.1.0,<2.0.0"
]

[project.urls]
"Homepage" = "https://github.com/ippo252525/pyneople"
"Bug Tracker" = "https://github.com/ippo252525/pyneople/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true